feat: Swift package for Apple platforms (Phase 5)
swift/: SPM package wrapping baedeker-ffi for macOS 13+ / iOS 16+:

- BaedekerModule (compilingWasm: / aotArtifact: / aotArtifact()
  serializer) and BaedekerInstance (call, callAsync serialized on a
  per-instance queue, registerHostFunction with Swift closures, fuelLimit)
- WasmValue enum with SIMD16<UInt8> v128; WasmValueType for signatures;
  BaedekerError mapping every status with the runtime's message
- Memory: readMemory/writeMemory (bounds-checked copies),
  withUnsafeMemoryBytes (zero-copy borrow, invalidation documented)
- Host closures via Unmanaged boxes + a C trampoline; thrown Swift
  errors trap the guest with the error description as the message
- build-xcframework.sh: baedeker-ffi for macOS (universal), iOS device,
  iOS simulator (universal) -> Vendor/CBaedeker.xcframework (gitignored)
- XCTest suite: compile/call/host-closure, AOT roundtrip, memory APIs,
  async call, error mapping, host failure
- macOS CI job: build-xcframework.sh + swift build + swift test

Also in this PR (FFI/CLI support for the package):
- baedeker_module_write_aot (C-side AOT serialize; artifact size query
  via null buffer)
- baedeker-cli compile <in.wasm> -o <out.bdkaot> — the build-time half
  of the iOS AOT workflow (documented in swift/README.md)
- BaedekerValue.tag is now a plain uint8_t carrying a BaedekerValueTag:
  fixed-underlying-type enums are C23-only, and a C-int enum field read
  3 bytes of undefined padding against Rust's 1-byte write (found by the
  Linux Swift harness). cbindgen post-process keeps both typedef
  branches enum-typed so Swift's Clang importer sees one candidate.
- build.rs post-processing documented; BaedekerValueTag force-included
  in the header for its constants

Validated end-to-end on Linux (Swift 6.1.2): compile -> AOT roundtrip
-> instantiate -> host closure -> call -> memory -> async -> error
mapping. Apple SPM path validated by the new macOS CI job.

+++ crates/baedeker-ffi/tests/ffi_roundtrip.rs
@@ -41,7 +41,7 @@ unsafe extern "C" fn host_add_ok(
     assert_eq!(n_results, 1);
     let args = unsafe { std::slice::from_raw_parts(args, n_args) };
     let (a, b) = unsafe { (args[0].data.i32_, args[1].data.i32_) };
-    unsafe { (*results).tag = BaedekerValueTag::I32 };
+    unsafe { (*results).tag = BaedekerValueTag::I32 as u8 };
     unsafe { (*results).data.i32_ = a + b };
     BaedekerStatus::Ok
 }
@@ -77,16 +77,16 @@ impl Instance {
     fn call_i32s(&self, name: &CStr, args: [i32; 2]) -> (BaedekerStatus, i32) {
         let c_args = [
             BaedekerValue {
-                tag: BaedekerValueTag::I32,
+                tag: BaedekerValueTag::I32 as u8,
                 data: BaedekerValueData { i32_: args[0] },
             },
             BaedekerValue {
-                tag: BaedekerValueTag::I32,
+                tag: BaedekerValueTag::I32 as u8,
                 data: BaedekerValueData { i32_: args[1] },
             },
         ];
         let mut out = [BaedekerValue {
-            tag: BaedekerValueTag::I32,
+            tag: BaedekerValueTag::I32 as u8,
             data: BaedekerValueData { i32_: 0 },
         }];
         let mut n_out = 0usize;
